+++ Server/package.json
@@ -16,6 +16,7 @@
     "botkit": "^4.15.0",
     "dotenv": "^16.4.5",
     "mongodb": "^6.9.0",
+    "openai": "^4.68.0",
     "socket.io": "^4.8.0"
   }
 }
